Add cards, section components and main content
Create the card component, which holds links to articles and other pages. Put all the data needed to generate a card component in a json file called data.json. Section then dynamically generates a section, with card components from the data.json which is handled by the MainSection.tsx component.